testsuite: move libutil tests to subdirectory
Instead of defining TEST_MAIN, move tests for optparse.c and nodeset.c
to separate source files in a test/ subdirectory.

The nodeset tests were accessing some internal constants and computing
sizes of internal objects.  To allow these tests to run from a different
module, added nodeattr_getattr () as a general accessor for internal
constants and sizes.
