+++ src/common/libutil/monotime.h
@@ -2,6 +2,7 @@
 #define _UTIL_MONOTIME_H
 
 #include <time.h>
+#include <stdbool.h>
 
 double monotime_since (struct timespec t0); /* milliseconds */
 void monotime (struct timespec *tp);
